+++ CHANGES.md
@@ -1,7 +1,7 @@
 
 # current
   - improved `make` (now O(log n))
-  - added `init`, `of_array`, `map`
+  - added `init`, `of_array`, `map`, `of_list`
 
 # 1.0.0
   - first release